requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.3.1
greenlet==3.0.1
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==13.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.14
toml==0.10.2
tornado==6.3.3
tzlocal==5.2
urllib3==2.0.7
youtube-dl==2021.12.17